requirements.txt
# Geospatial
geojson==3.1.0
mapbox-vector-tile==1.2.1
geopy==2.4.1
vt2geojson==0.2.1
shapely==2.0.2  # Libreria per operazioni geometriche
folium==0.19.4

# Computer Vision & ML
opencv-python-headless==4.9.0.80
tensorflow==2.16.1
pandas==2.2.3

# Utilities & Networking
requests==2.31.0
protobuf==3.20.3

# Environment Variables
python-dotenv==1.0.0

# MLOps
mlflow==2.21.0

# Load pre-trained models
transformers==4.50.3

# Hugging face's library for load local dataset
datasets==2.19.1